% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/IRTShiny.R
\name{startIRT}
\alias{startIRT}
\title{This function will start IRTShiny}
\usage{
startIRT()
}
\value{
Nothing
}
\description{
An interactive Shiny application for running a IRT analysis.
}
\details{
Start IRTShiny

This starts the Shiny application IRT.
}
\examples{
\dontrun{
library(shiny)
startIRT()
}
}
\keyword{IRT}

